requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
google-generativeai
